+++ components/drawer/style/index.scss
@@ -1,5 +1,4 @@
 @import '../../style/themes/default';
 @import './component';
-
 @import '~zarm/lib/popup/style/index.css';
 @import '~zarm/lib/mask/style/index.css';
